Refactor agent logic into its own module
